@@ -1,12 +1,13 @@
 
-namespace Entity
+namespace Model
 {
-    public class Tag
+    public class Category
     {
         public int Id { get; set; }
         public string Name { get; set; }
-   
     }
 }
+
+
 
 
