@@ -1,13 +1,12 @@
 
-namespace Entity
+namespace Model 
 {
-    public class Category
+    public class Tag
     {
         public int Id { get; set; }
         public string Name { get; set; }
+        public int ProductId { get; set; }
     }
 }
-
-
 
 
